+++ plugins/toubiz/index.js
@@ -11,7 +11,7 @@ export function createPlugin() {
     metadata: {
       pluginType: 'toubiz',
       name: 'Toubiz Event Management',
-      version: '1.0.0',
+      version: '1.1.0',
     },
 
     async initialize() {},